app/dependencies.py
from fastapi import Depends, HTTPException, status, requests
from jose import JWTError
from sqlalchemy.orm import Session
import requests
from .services.auth_service import decode_access_token
from .db import get_db
from .services.auth_service import oauth2_scheme

def get_current_user_id(token: str = Depends(oauth2_scheme)) -> int:
    try:
        payload = decode_access_token(token)
        user_id: int = payload.get("user_id")
        if user_id is None:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid token",
                headers={"WWW-Authenticate": "Bearer"},
            )
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token",
            headers={"WWW-Authenticate": "Bearer"},
        )

    return user_id